+++ src/components/canary/atoms/select/index.ts
@@ -1,0 +1,12 @@
+export {
+  ArdaSelect,
+  Select,
+  SelectContent,
+  SelectGroup,
+  SelectItem,
+  SelectLabel,
+  SelectSeparator,
+  SelectTrigger,
+  SelectValue,
+} from './select';
+export type { ArdaSelectProps } from './select';